Moves subtitlesetting file to respect convention
